correctly use the default HTTP agent when bypass proxy is set
passing { HttpOptions: { agent: undefined } } causes the AWS sdk
to not use a http agent at all, which is very undesirable (no
connection pooling).

Move the proxy configuration into a new singleton so that it can
be overriden on a lib by lib basis rather than in the global
AWS configuration.  And only pass the proxyAgent setting for dynamo
if it's explictly specified.